+++ clx_budget_analysis_report/models/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo, CLx Media
 # See LICENSE file for full copyright & licensing details.
-from . import models
-from . import report
-
+from . import sale_subscription_line